Add restaurant open and close times to reservation card
Modified the restaurant page and ReservationCard component in the code base to include opening and closing times for the restaurant. This involved fetching these times from the backend and passing them to the ReservationCard as props. A new file "times.ts" was created to store an array of possible time slots and a method to filter this array was added to the ReservationCard to only display the times within the working hours.