Add availability route for restaurant API
This commit introduces a new route to the restaurant API for checking availability. It uses 'GET' method and accepts 'day', 'time', and 'partySize' parameters. For now, it just checks and returns search times, but will be built out to return all availability data needed by the frontend. It also returns an error message when any of the 3 parameters are missing.