Uncomment table creation in seed route
The code for table creation in the seed route file was previously commented out. This commit has uncommented these lines enabling the application to create tables based on the given data. This action allows us to visit the seed endpoint to seed tables with the specific arrangement of seats for the 'vivaanId' restaurant. We need these as sample data to develop and test the full availability system with.